Add support for skip pattern
