+++ tck/pom.xml
@@ -52,6 +52,12 @@
             <scope>provided</scope>
         </dependency>
         <dependency>
+            <groupId>org.eclipse.microprofile.config</groupId>
+            <artifactId>microprofile-config-api</artifactId>
+            <version>${version.mp.config}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
             <!-- actually only referenced in JavaDoc -->
             <groupId>org.apache.geronimo.specs</groupId>
             <artifactId>geronimo-annotation_1.2_spec</artifactId>
